add command validation by annotation

+++ src/main/java/superapp/logic/mongo/MiniAppCommandManagerMongoDB.java
@@ -11,13 +11,11 @@ import org.springframework.jms.core.JmsTemplate;
 import org.springframework.stereotype.Service;
 import superapp.data.*;
 import superapp.logic.MiniAppCommandWithAsyncSupport;
-import superapp.logic.boundaries.InvokedBy;
-import superapp.logic.boundaries.TargetObject;
+import superapp.logic.boundaries.*;
 import superapp.logic.utils.convertors.CommandConvertor;
-import superapp.logic.utils.convertors.ConvertIdsHelper;
+import superapp.logic.utils.validators.BoundaryValidator;
+import superapp.logic.utils.validators.EntitiesValidator;
 import superapp.miniapps.MiniAppNames;
-import superapp.logic.boundaries.CommandId;
-import superapp.logic.boundaries.MiniAppCommandBoundary;
 import superapp.miniapps.command.CommandInvoker;
 import superapp.miniapps.command.InvalidCommand;
 import superapp.miniapps.command.MiniAppsCommand;
@@ -28,27 +26,31 @@ import java.util.*;
 @Service
 public class MiniAppCommandManagerMongoDB implements MiniAppCommandWithAsyncSupport {
 
+    private String springApplicationName;
     private ObjectMapper jackson;
     private JmsTemplate jmsTemplate;
     private final MiniAppCommandCrud miniAppCommandCrud;
-    private final UserCrud userCrud;
-    private final ObjectCrud objectCrud;
     private final CommandInvoker commandInvoker;
     private final RBAC accessControl;
-    private String springApplicationName;
-
+    private final BoundaryValidator boundaryValidator;
+    private final EntitiesValidator entitiesValidator;
     private final CommandConvertor commandConvertor;
 
+
     @Autowired
-    public MiniAppCommandManagerMongoDB(MiniAppCommandCrud miniAppCommandCrud, UserCrud userCrud,
-                                        ObjectCrud objectCrud, CommandInvoker commandInvoker,
-                                        RBAC accessControl, CommandConvertor commandConvertor) {
+    public MiniAppCommandManagerMongoDB(MiniAppCommandCrud miniAppCommandCrud,
+                                        CommandInvoker commandInvoker,
+                                        CommandConvertor commandConvertor,
+                                        BoundaryValidator boundaryValidator,
+                                        EntitiesValidator entitiesValidator,
+                                        RBAC accessControl) {
+
         this.miniAppCommandCrud = miniAppCommandCrud;
-        this.userCrud = userCrud;
-        this.objectCrud = objectCrud;
         this.commandInvoker = commandInvoker;
-        this.accessControl = accessControl;
         this.commandConvertor = commandConvertor;
+        this.accessControl = accessControl;
+        this.boundaryValidator = boundaryValidator;
+        this.entitiesValidator = entitiesValidator;
     }
 
     @Autowired
@@ -74,92 +76,35 @@ public class MiniAppCommandManagerMongoDB implements MiniAppCommandWithAsyncSupp
     }
 
 
-    private boolean isValidInvokedBy(InvokedBy invokedBy) {
-        // check if invokedBy is null
-        if (invokedBy == null)
-            return false;
-
-        // check if invokedBy fields are null
-        if (invokedBy.getUserId() == null)
-            return false;
-
-        return invokedBy.getUserId().getEmail() != null &&
-                !invokedBy.getUserId().getEmail().isEmpty() &&
-                invokedBy.getUserId().getSuperapp() != null &&
-                !invokedBy.getUserId().getSuperapp().isEmpty();
-    }
-
-    private boolean isValidTargetObject(TargetObject targetObject) {
-        // check if targetObject is null
-        if (targetObject == null) return false;
-
-        // check if targetObject fields are null
-
-        // check if targetObject fields are null
-        if (targetObject.getObjectId() == null) return false;
-
-        return targetObject.getObjectId().getInternalObjectId() != null &&
-                !targetObject.getObjectId().getInternalObjectId().isEmpty() &&
-                targetObject.getObjectId().getSuperapp() != null &&
-                !targetObject.getObjectId().getSuperapp().isEmpty();
-    }
-
-
+    /**
+     * This methode execute the commands by commandInvoker
+     * @param commandBoundary
+     * @return Object command result
+     */
     @Override
     public Object invokeCommand(MiniAppCommandBoundary commandBoundary) {
 
-        InvokedBy invokedBy = commandBoundary.getInvokedBy();
-        if (!isValidInvokedBy(invokedBy))
-            throw new BadRequestException("invoked by is not valid");
+        // validation
+        this.boundaryValidator.validateCommandBoundary(commandBoundary);
 
-        String userId = ConvertIdsHelper.concatenateIds(new String[]{
-                invokedBy.getUserId().getSuperapp(), invokedBy.getUserId().getEmail()});
-
-        UserEntity user = this.userCrud
-                .findById(userId)
-                .orElseThrow(() -> new NotFoundException("user " + userId + " not found in database"));
-
-        // check permission
-        if (!accessControl.hasPermission(userId, "invokeCommand"))
-            throw new UnauthorizedRequestException("user " + userId + " with role " + user.getRole() + " has no permission to invokeCommand");
-
-        // validate target object
-        if (!isValidTargetObject(commandBoundary.getTargetObject()))
-            throw new BadRequestException("target object is not valid");
+        // validate that user exist and retrieve the user from database
+        UserId userId = commandBoundary.getInvokedBy().getUserId();
+        UserEntity userEntity = this.entitiesValidator
+                .validateExistingUser(userId.getSuperapp(), userId.getEmail());
 
 
+        // validate target object exist in database
+        ObjectId objectId = commandBoundary.getTargetObject().getObjectId();
+        this.entitiesValidator
+                .validateExistingObject(objectId.getSuperapp(), objectId.getInternalObjectId());
 
-        SuperAppObjectEntity targetObjectEntity;
-        String objectId = this.commandConvertor.targetObjToEntity(commandBoundary.getTargetObject());
-
-        // validate object exist or is a default object
-        String idForCommandWithoutTarget = "a114f4a4-2be4-43ad-90f7-9822e8867d5e";
-        if (this.objectCrud.existsById(objectId)) {
-
-            targetObjectEntity = this.objectCrud
-                    .findById(objectId)
-                    .orElseThrow(() -> new NotFoundException("target object not exist in data base"));
-
-            // validate object exist is not active:false
-            if (!targetObjectEntity.getActive())
-                throw new NotFoundException(" target object id " + targetObjectEntity.getObjectId() + "not found - active:false");
-
-        }
-        else if (!commandBoundary.getTargetObject().getObjectId().getInternalObjectId().equals(idForCommandWithoutTarget)
-                || !commandBoundary.getTargetObject().getObjectId().getSuperapp().equals(this.springApplicationName)) {
-            // TODO: create default object for all commands, instead idForCommandWithoutTarget
-            throw new NotFoundException("target object not exist in data base");
-        }
-
-
-        if (commandBoundary.getCommand() == null)
-            throw new BadRequestException("command cant be null");
+        checkPermission(userEntity.getUserID(), "invokeCommand");
 
         // set command id and invocation timestamp
         CommandId commandId = commandBoundary
-                .getCommandId()
-                .setInternalCommandId(UUID.randomUUID().toString())
-                .setSuperapp(springApplicationName);
+                                        .getCommandId()
+                                        .setInternalCommandId(UUID.randomUUID().toString())
+                                        .setSuperapp(this.springApplicationName);
 
         commandBoundary
                 .setCommandId(commandId)
@@ -169,34 +114,21 @@ public class MiniAppCommandManagerMongoDB implements MiniAppCommandWithAsyncSupp
         Object resultObjectOfCommand;
 
         MiniAppNames miniappName = MiniAppNames.strToMiniAppName(commandBoundary.getCommandId().getMiniapp());
-
         MiniAppsCommand.commands commandsToExecute = MiniAppsCommand.strToCommand(commandBoundary.getCommand());
 
-
-        if (miniappName.equals(MiniAppNames.UNKNOWN)) {
-            // TODO : is align with the spec?
-            resultObjectOfCommand =
-                    new InvalidCommand(miniappName + " miniapp name: "
-                            + commandBoundary.getCommandId().getMiniapp() + " not supported");
-
-        }
-        else if (commandsToExecute.equals(MiniAppsCommand.commands.UNKNOWN_COMMAND)) {
-            resultObjectOfCommand =
-                    new InvalidCommand(miniappName + " miniapp command: "
-                            + commandBoundary.getCommand() + " not supported");
-
-        }
-        else {
+        if (miniappName.equals(MiniAppNames.UNKNOWN)) // TODO : is align with the spec?
+            resultObjectOfCommand = new InvalidCommand(miniappName + " miniapp name: "
+                                                        + commandBoundary.getCommandId().getMiniapp() + " not supported");
+        else if (commandsToExecute.equals(MiniAppsCommand.commands.UNKNOWN_COMMAND))
+            resultObjectOfCommand = new InvalidCommand(miniappName + " miniapp command: "
+                                                    + commandBoundary.getCommand() + " not supported");
+        else
             resultObjectOfCommand = commandInvoker
-                    .create(commandsToExecute)
-                    .execute(commandBoundary);
-
-        }
-
+                                        .create(commandsToExecute)
+                                        .execute(commandBoundary);
 
         // convert to entity
-        //MiniAppCommandEntity commandEntity = convertToEntity(commandBoundary);
-        MiniAppCommandEntity commandEntity = commandConvertor.toEntity(commandBoundary);
+        MiniAppCommandEntity commandEntity = this.commandConvertor.toEntity(commandBoundary);
 
         // save command to database
         this.miniAppCommandCrud.save(commandEntity);
@@ -209,21 +141,23 @@ public class MiniAppCommandManagerMongoDB implements MiniAppCommandWithAsyncSupp
     }
 
 
+    /**
+     * This method retrieve all commands history of miniapp
+     *
+     * @param userSuperapp String
+     * @param userEmail String
+     * @param size int
+     * @param page int
+     * @return List MiniAppCommandBoundary
+     */
     @Override
     public List<MiniAppCommandBoundary> getAllMiniAppCommands(String miniAppName, String userSuperapp, String userEmail,
                                                               int size, int page) {
 
+        // validate that user exist and retrieve the user from database
+        UserEntity userEntity = this.entitiesValidator.validateExistingUser(userSuperapp, userEmail);
 
-        String userId = ConvertIdsHelper.concatenateIds(new String[]{ userSuperapp, userEmail});
-
-        // validate that user exist
-        this.userCrud
-                .findById(userId)
-                .orElseThrow(() -> new NotFoundException("user " + userId + " not found in database"));
-
-        // check role permission
-        if (!accessControl.hasPermission(userId, "getAllMiniAppCommands"))
-            throw new UnauthorizedRequestException("user " + userId + " has no permission to getAllMiniAppCommands");
+        this.checkPermission(userEntity.getUserID(), "getAllMiniAppCommands");
 
         // Check if the miniApp name is valid
         if (MiniAppNames.strToMiniAppName(miniAppName).equals(MiniAppNames.UNKNOWN))
@@ -239,19 +173,22 @@ public class MiniAppCommandManagerMongoDB implements MiniAppCommandWithAsyncSupp
                 .toList();
     }
 
+    /**
+     * This method retrieve all commands history
+     *
+     * @param userSuperapp String
+     * @param userEmail String
+     * @param size int
+     * @param page int
+     * @return List MiniAppCommandBoundary
+     */
     @Override
     public List<MiniAppCommandBoundary> getAllCommands(String userSuperapp, String userEmail, int size, int page) {
 
-        String userId = ConvertIdsHelper.concatenateIds(new String[]{ userSuperapp, userEmail});
+        // validate that user exist and retrieve the user from database
+        UserEntity userEntity = this.entitiesValidator.validateExistingUser(userSuperapp, userEmail);
 
-        // validate that user exist
-        this.userCrud
-                .findById(userId)
-                .orElseThrow(() -> new NotFoundException("user " + userId + " not found in database"));
-
-        // check role permission
-        if (!accessControl.hasPermission(userId, "getAllCommands"))
-            throw new UnauthorizedRequestException("user " + userId + " has no permission to getAllCommands");
+        this.checkPermission(userEntity.getUserID(), "getAllCommands");
 
 
         PageRequest pageRequest = PageRequest
@@ -265,18 +202,31 @@ public class MiniAppCommandManagerMongoDB implements MiniAppCommandWithAsyncSupp
     }
 
 
+    /**
+     * This method delete all commands, only admin has permission
+     *
+     * @param userSuperapp String
+     * @param userEmail String
+     */
     @Override
     public void deleteAllCommands(String userSuperapp, String userEmail) {
 
-        String userId = ConvertIdsHelper.concatenateIds(new String[]{ userSuperapp, userEmail});
-        // check role permission
-        if (!accessControl.hasPermission(userId, "deleteAllCommands"))
-            throw new UnauthorizedRequestException("user " + userId + " has no permission to deleteAllCommands");
+        // validate that user exist and retrieve the user from database
+        UserEntity userEntity = this.entitiesValidator.validateExistingUser(userSuperapp, userEmail);
+
+        this.checkPermission(userEntity.getUserID(), "deleteAllCommands");
 
         this.miniAppCommandCrud.deleteAll();
     }
 
+    private void checkPermission(String userId, String operationName) {
+        // check role permission
+        if (!this.accessControl.hasPermission(userId, operationName))
+            throw new UnauthorizedRequestException("User " + userId + " has no permission to " + operationName);
+    }
 
+
+    /**** Support Async *****/
     @Override
     public Object asyncHandle(MiniAppCommandBoundary miniAppCommandBoundary) {
 
@@ -337,27 +287,26 @@ public class MiniAppCommandManagerMongoDB implements MiniAppCommandWithAsyncSupp
     }
 
 
-
     /**** Deprecated methods *****/
     @Deprecated
     @Override
     public List<MiniAppCommandBoundary> getAllMiniAppCommands(String miniAppName) {
         throw new DeprecatedRequestException("cannot enter a deprecated function");
-//
-//        // Check if the miniApp name is valid
-//        if (!validMiniAppName(miniAppName)) throw new BadRequestException("MiniApp name is not valid");
-//
-//        // Create a list of commands
-//        List<MiniAppCommandBoundary> commandBoundaryList = new ArrayList<>();
-//
-//        // Get all commands from the database and filter by miniApp name
-//        this.miniAppCommandCrud.findAll().forEach(commandEntity -> {
-//            if (commandEntity.getCommandId().contains(miniAppName)) {
-//                commandBoundaryList.add(convertToBoundary(commandEntity));
-//            }
-//        });
-//
-//        return commandBoundaryList;
+        //
+        //        // Check if the miniApp name is valid
+        //        if (!validMiniAppName(miniAppName)) throw new BadRequestException("MiniApp name is not valid");
+        //
+        //        // Create a list of commands
+        //        List<MiniAppCommandBoundary> commandBoundaryList = new ArrayList<>();
+        //
+        //        // Get all commands from the database and filter by miniApp name
+        //        this.miniAppCommandCrud.findAll().forEach(commandEntity -> {
+        //            if (commandEntity.getCommandId().contains(miniAppName)) {
+        //                commandBoundaryList.add(convertToBoundary(commandEntity));
+        //            }
+        //        });
+        //
+        //        return commandBoundaryList;
     }
 
     @Override
